bot refactor, wip front
