+++ TODO.md
@@ -1,9 +1,6 @@
 
 ## Por hacer:
 
-- Agregar un formulario web, quiza en react, para mandar el post a **webservice** que se guarda en la bd sqlite y el bot lea la data de esa bd.
-
-
 - Algoritmo de respuesta para challenges aleatorios dependiendo de: 
 
     1. **level**: ["easy", "medium", "hard"]
