+++ botservice/messages/status.go
@@ -1,4 +1,4 @@
-package status
+package messages
 
 import (
 	"log"
@@ -7,6 +7,13 @@ import (
 )
 
 func SetStatus(discord *discordgo.Session, ready *discordgo.Ready) {
+	// TODO: update here when fixed by Discord
+	//_, err := discord.UserUpdateStatus(discordgo.StatusDoNotDisturb)
+	err := discord.UpdateStatusComplex(discordgo.UpdateStatusData{Status: "dnd"})
+	if err != nil {
+		log.Println(err.Error())
+	}
+
 	_ = discord.UpdateGameStatus(1, ".go Googleando")
 	servers := discord.State.Guilds
 	log.Printf("Bot iniciado en %d servers", len(servers))
